[C-4419] Simplify SSR mobile track page
