+++ packages/harmony/src/components/text/Text.tsx
@@ -19,7 +19,7 @@ export const Text = forwardRef(
       variant: propVariant,
       strength: strengthProp,
       size: sizeProp,
-      color,
+      color: colorProp,
       shadow,
       tag,
       asChild,
@@ -34,6 +34,7 @@ export const Text = forwardRef(
     const variant = propVariant ?? contextVariant ?? 'body'
     const strength = strengthProp ?? (contextVariant ? undefined : 'default')
     const size = sizeProp ?? (contextVariant ? undefined : 'm')
+    const color = colorProp ?? (contextVariant ? undefined : 'default')
 
     const variantConfig = variant && variantStylesMap[variant]
     const css = {
