--- packages/web/src/pages/server-track-page/ServerTrackPage.tsx
@@ -1,14 +1,0 @@
-import { ServerTrackPage as MobilePage } from './mobile/ServerTrackPage'
-
-export type ServerTrackPageProps = {
-  track: any
-  isMobile: boolean
-}
-
-export const ServerTrackPage = ({
-  isMobile,
-  ...other
-}: ServerTrackPageProps) => {
-  // @ts-ignore
-  return isMobile ? <MobilePage {...other} /> : null
-}
