--- packages/web/src/pages/server-track-page/mobile/ServerActionButtonRow.tsx
@@ -1,14 +1,0 @@
-import { Flex } from '@audius/harmony'
-
-export const ServerActionButtonRow = () => {
-  return (
-    <Flex
-      h={42.5}
-      direction='row'
-      alignItems='center'
-      justifyContent='center'
-      gap='xl'
-      mt='m'
-    />
-  )
-}
